# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for tms",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/translations": {
            "get": {
                "tags": [
                    "translations"
                ],
                "summary": "Get all translations",
                "description": "Returns a list of translations",
                "operationId": "c902e5899ce8b1bb38b46fccf05bc0e2",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "translations"
                ],
                "summary": "Create a new translation",
                "operationId": "e7f924a9baac9399ce933fa98ed6b00b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale",
                                    "key",
                                    "content",
                                    "tags"
                                ],
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "example": "eng"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "welcome_message"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "{\\`en\\`:\\`Welcome\\`,\\`es\\`:\\`Bienvenido\\`}"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "greeating",
                                            "welcome"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "locale": {
                                            "type": "string",
                                            "example": "eng"
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "welcome_message"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "{\\`en\\`:\\`Welcome\\`,\\`es\\`:\\`Bienvenido\\`}"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "greeating",
                                                "welcome"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/translations/search": {
            "get": {
                "tags": [
                    "translations"
                ],
                "summary": "Search translations",
                "operationId": "f022d0e019f5aeba58ffcc9f38a6b96e",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query string",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "locale": {
                                                "type": "string",
                                                "example": "eng"
                                            },
                                            "key": {
                                                "type": "string",
                                                "example": "welcome_message"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "{\\`en\\`:\\`Welcome\\`,\\`es\\`:\\`Bienvenido\\`}"
                                            },
                                            "tags": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "greeating",
                                                    "welcome"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/tags/{tagName}": {
            "get": {
                "tags": [
                    "translations"
                ],
                "summary": "Get translations by tag name",
                "operationId": "857d1548ac310732521ce5ef7ca1cfee",
                "parameters": [
                    {
                        "name": "tagName",
                        "in": "path",
                        "description": "Tag name to filter translations",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "locale": {
                                                "type": "string",
                                                "example": "eng"
                                            },
                                            "key": {
                                                "type": "string",
                                                "example": "welcome_message"
                                            },
                                            "tags": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "greeating",
                                                    "welcome"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/{id}/assign-tags": {
            "post": {
                "tags": [
                    "translations"
                ],
                "summary": "Assign tags to a translation",
                "operationId": "623d8d4175f0088f0fec230ea10b3907",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "greeting",
                                            "welcome"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tags assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tags assigned successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/export": {
            "get": {
                "tags": [
                    "translations"
                ],
                "summary": "Export all translations",
                "operationId": "02d8ad9e5cd98059ea33d8f4bb6e3494",
                "responses": {
                    "200": {
                        "description": "Exported translations"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "The **Authorization** header using the Bearer scheme. Example: **Bearer {token}**",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": {
        "sanctum": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    }
}